pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "playwright-test-generator"
version = "0.1.0"
description = "AI-powered Playwright test generator using local LLM models via Ollama"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.3.0",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ollama>=0.1.0",
    "streamlit>=1.32.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "plotly>=6.7.0",
    "pandas>=2.3.3",
]
[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.15.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "types-requests>=2.31.0",
    "pre-commit>=4.5.1",
    "pandas-stubs>=3.0.0.260204",
]
[tool.hatch.build.targets.wheel]
packages = ["src", "cli"]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = ["generated_tests"]
[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

ignore = ["E501", "B027"]
[tool.ruff.lint.isort]
known-first-party = ["src", "cli"]
[tool.mypy]
python_version = "3.13"
exclude = ["generated_tests"]
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "plotly.*"
ignore_missing_imports = true
